Перенес коннект из withCommentHOC в AddCommentForm
