+++ src/hocs/with-comment/with-comment.js
@@ -1,10 +1,6 @@
 import React, {PureComponent} from 'react';
-import {connect} from 'react-redux';
-import {AppRoutes, FieldsIds} from '../../consts/consts.js';
-import {Operation as CommentsOperation, ActionCreator as CommentsActionCreator} from '../../reducer/data/comments-data/comments-data.js';
+import {FieldsIds} from '../../consts/consts.js';
 import PropTypes from 'prop-types';
-import {getAuthorizationStatusBoolean, getIsPostCommentHasError, getIsPostCommentInProgress} from '../../reducer/selectors.js';
-import {Redirect} from 'react-router-dom';
 import {isValidField} from '../../utils/filters.js';
 
 const withComment = (Component) => {
@@ -18,7 +14,6 @@ const withComment = (Component) => {
       };
       this._handleRadioChange = this._handleRadioChange.bind(this);
       this._handleTextAreaChange = this._handleTextAreaChange.bind(this);
-      this._handleFormSubmit = this._handleFormSubmit.bind(this);
     }
 
     _handleRadioChange(rating) {
@@ -27,25 +22,6 @@ const withComment = (Component) => {
 
     _handleTextAreaChange(comment) {
       this.setState({comment});
-    }
-
-    _handleFormSubmit(evt) {
-      evt.preventDefault();
-      const {rating, comment} = this.state;
-      const {movieId, postComment} = this.props;
-      postComment({
-        movieId,
-        rating,
-        comment,
-      });
-    }
-
-    componentDidUpdate(prevProps) {
-      if (prevProps.postCommentInProgress !== this.props.postCommentInProgress && !this.props.postCommentError) {
-        this.setState({
-          isNeedRedirect: true,
-        });
-      }
     }
 
     componentWillUnmount() {
@@ -58,26 +34,13 @@ const withComment = (Component) => {
 
     render() {
       const {comment, rating} = this.state;
-      const {isAuthorized, postCommentInProgress, postCommentError} = this.props;
       const isFormValid = isValidField(FieldsIds.RATING_FIELD_ID, this.state.rating) && isValidField(FieldsIds.COMMENTS_FIELD_ID, this.state.comment);
-      if (!isAuthorized) {
-        return (
-          <Redirect to={AppRoutes.LOGIN_PAGE} />
-        );
-      }
-      if (this.state.isNeedRedirect) {
-        return (
-          <Redirect to={`${AppRoutes.FILM_PAGE}/${this.props.movieId}`} />
-        );
-      }
+
       return (
         <Component
           {...this.props}
           onTextAreaChange={this._handleTextAreaChange}
           onRadioChange={this._handleRadioChange}
-          onFormSubmit={this._handleFormSubmit}
-          postCommentInProgress={postCommentInProgress}
-          postCommentError={postCommentError}
           comment={comment}
           rating={rating}
           isFormValid={isFormValid}
@@ -86,30 +49,13 @@ const withComment = (Component) => {
     }
   }
 
-  const mapStateToProps = (state) => {
-    return {
-      isAuthorized: getAuthorizationStatusBoolean(state),
-      postCommentInProgress: getIsPostCommentInProgress(state),
-      postCommentError: getIsPostCommentHasError(state),
-    };
-  };
-
-  const mapDispatchToProps = (dispatch) => ({
-    postComment(commentData) {
-      dispatch(CommentsActionCreator.postComment());
-      dispatch(CommentsOperation.postCommentData(commentData));
-    },
-  });
-
   WithCommentHOC.propTypes = {
     movieId: PropTypes.number.isRequired,
     postComment: PropTypes.func.isRequired,
     isAuthorized: PropTypes.bool.isRequired,
-    postCommentInProgress: PropTypes.bool.isRequired,
-    postCommentError: PropTypes.bool.isRequired,
   };
 
-  return connect(mapStateToProps, mapDispatchToProps)(WithCommentHOC);
+  return WithCommentHOC;
 };
 
 export default withComment;
